+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Unreleased
+
+## 1.0.9
 - Control collision and race-condition for spread keys
 - Merged `fnjson` and `fn` functionalities